rename gl_constants -> constants & remove unused values (add back as needed)
